requirements.txt
redis==3.5.3
skale.py==5.6dev0
